backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
pydantic-settings==2.4.0
boto3==1.34.162
redis==5.0.7
rq==1.16.2
sqlalchemy==2.0.32
pdfplumber==0.11.4
pytesseract==0.3.13
Pillow==10.4.0
python-slugify==8.0.4
httpx==0.27.0
tenacity==8.5.0
